fix: getCurrentSemester 학기랑 년도 순서 바뀌어 있던 거 고침 + 기타 항목 신청 리스트 GET API 구현 성공

+++ src/main/java/com/csee/swplus/mileage/etcSubitem/dto/EtcSubitemResponseDto.java
@@ -8,9 +8,9 @@ import java.util.List;
 @Data
 @AllArgsConstructor
 public class EtcSubitemResponseDto {    // 관리자가 신청 받기를 허용한 (subitem 테이블의 student_input = 'Y') 기타 항목 (subitem) 전용 DTO
-    private Long subitemId;
+    private int subitemId;
     private String subitemName;
-    private Long categoryId;
+    private int categoryId;
     private String categoryName;
     private String semester;
 }
